Rename svg.cpp to render.c
It is plain C not a C++ code. Also use name better describing what
is a purpose of this file.
