+++ src/t_render.c
@@ -1,23 +1,23 @@
 #include <string.h>
-#include "svg.hpp"
+#include "render.h"
 
 #include "unity_fixture.h"
 
-TEST_GROUP(SvgGroup);
+TEST_GROUP(RenderGroup);
 
-TEST_SETUP(SvgGroup) {}
+TEST_SETUP(RenderGroup) {}
 
-TEST_TEAR_DOWN(SvgGroup) {}
+TEST_TEAR_DOWN(RenderGroup) {}
 
-TEST(SvgGroup, ParseSVGTest)
+TEST(RenderGroup, ParseSVGTest)
 {
 	const char *whiterect =
 		"<svg width=\"640\" height=\"150\" version=\"1.1\"><rect "
 		"x=\"1\" y=\"0\" width=\"638\" height=\"150\" ry=\"0\" "
 		"fill=\"#ffffff\"/></svg>";
 
-	unsigned char *buffer = (unsigned char *)malloc(BITMAP_BYTES);
-	svgtorgb565(whiterect, strlen(whiterect), (char *)buffer);
+	void *buffer = malloc(BITMAP_BYTES);
+	svgtobgr565(whiterect, strlen(whiterect), buffer);
 
 	TEST_ASSERT_EQUAL(0x00, *(buffer));
 	TEST_ASSERT_EQUAL(0xff, *(buffer + 2));
@@ -28,16 +28,16 @@ TEST(SvgGroup, ParseSVGTest)
 }
 
 
-TEST_GROUP_RUNNER(SvgGroup)
+TEST_GROUP_RUNNER(RenderGroup)
 { /* Add a line below for each unit test */
 
-	RUN_TEST_CASE(SvgGroup, ParseSVGTest);
+	RUN_TEST_CASE(RenderGroup, ParseSVGTest);
 }
 
 static void
 runAllTests()
 {
-	RUN_TEST_GROUP(SvgGroup);
+	RUN_TEST_GROUP(RenderGroup);
 }
 
 int
